+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/MetroCompilerTest.kt
@@ -231,6 +231,9 @@ abstract class MetroCompilerTest {
               MetroOption.ENABLE_DAGGER_ANVIL_INTEROP -> {
                 processor.option(entry.raw.cliOption, enableDaggerAnvilInterop)
               }
+              MetroOption.ENABLE_INJECT_CONSTRUCTOR_HINTS -> {
+                processor.option(entry.raw.cliOption, enableInjectConstructorHints)
+              }
             }
           yield(option)
         }